fix #58 empty error state on action service failure
If an action service method call resulted in an error an empty object
is committed to the `set*Error` mutation rather than the error details.
This was an unintended result of `Object.assign({}, error)` which
always returned an empty object.

The mutation is responsible for serializing the error, via the
‘serialize-error’ package, and updating the `errorOn*` state so passing
the rejected error object as the mutation argument should work.