Consistently update objects
The updateItem mutation, the commitCopy, mutation and the model now all use the same `updateOriginal` function to update items in the store.

I’ve also updated the `updateOriginal` function to use Vue.set only when the property doesn’t already exist.  If it’s an existing property, it will use the equals operator to overwrite it.  This should avoid breaking any reactivity.  fingers crossed