Turn temps off in find getter by default
As pointed out by @t2t2 in #216, it’s more intuitive to opt in to the temps feature than to include temps by default.  It leads to some confusing behavior when new records just start appearing.

It’s also a lesser use case to need temp records in an app.