Refactor & simplify the find getter
The find getter will never use the FETCH_PARAMS, so it doesn’t need to use the `getParams` function at all.  This means the `getParams` function doesn’t need any branching logic centered around the find getter, so it has been simplified.

These changes also make it so that if FETCH_PARAMS and PARAMS are both provided, the find getter will NOT use the internal pagination feature, and instead will behave the same as it does in feathers-vued@1.x.