Fix watcher for makeGetMixin
The watcher in the makeGetMixin was calling the GET_ACTION with the params, which was effectively setting the params as the id for the query.  Naturally, some really weird errors were showing up on the API server.  This makes sure the watcher doesn’t call the GET_ACTION with any params, allowing the GET_ACTION to perform its param-choosing logic.