fix: assure only one clone ever exists in memory
Clones are now handled in the same way that instances are.  If you call clone twice on the same object, the return value is always the same clone.  This assures that only one clone ever exists in memory, creating a very consistent API.  (Of course, more clones can always be stored anywhere in an app by manual use of a cloning utility like fast-copy)