fix: Fix memory leak fast-copying nested instance
Ok, it’s not a real memory leak, but we don’t want extra instances in memory.  This makes sure that fastCopy is never used on the first level of a model instance.  (It works fine on nested model instances, because it always calls new prop.constructor on nested objects)

It includes lots of tests for cloning and commiting.