Fix ts config
Recently a change was made to add `declaration: true` to the `tsconfig.json`.  This option generates the corresponding `d.ts` typing files.  The `allowJs` option cannot be used with this new option, so it is being remove.  We don’t actually need to directly compile js files, anyway, so this should affect nothing.