fix check for temp record
Checking for the presence of the idField on a record is a better indicator of if the record is a temp.  It beats checking for the lack of an `__id` field, because in Vue it’s sometimes a pain in the butt to get rid of properties thanks to using getters and setters for reactivity. This won’t be necessary once 3.0 ships with proxy-based reactivity.