test(package.json): Fix glob pattern in mocha test script
The test script in the package.json should have the glob pattern wrapped in single quotes like mentioned here:
https://mochajs.org/#the-test-directory

"You should always quote your globs in npm scripts. If you use double quotes, it's the shell on UNIX that will expand the glob. On the other hand, if you use single quotes, the node-glob module will handle its expansion."

If not, in some environments it will not take the files in the root directory into accounts.