Compare both model service path and user service path using the shortName format.
This addresses an issue where models defined with a service path containing a leading slash('/') won't cause a mismatch between the userService path.

Defining a leading slash in the userService path will cause an issue during the idField fetch on setup. This should be the easiest possible solution without much risk of any unexpected mismatches.