Set isDirty to false on reset()
From a UX perspective, this makes sense to me. If I'm resetting a clone, then there are no longer any changes present, and therefore the form is..."clean". I'm usually using a "Save" and "Reset" button which are set to :disabled="!isValid && !isDirty". In this case it would be nice to reset isDirty so that the reset button can be disabled when no changes are present.