Better check for detecting instances in getters
The isInstance check was only using instanceof, which wasn’t working well for proxy objects for some reason.  I assumed someVueProxyInstance instanceof MyModel would work, but it seems to not be the case.  This adds a fallback to check for the item’s constructor.idField, which is both unique and required for FeathersVuex Model classes