fix: removes legacy createElement in favor h
The `@vue/composition-api` package is now stable and createElement was replaced with h. 

The current implementation is throws vite errors (especially when building in cloud services like Cloudflare Pages for example).

This PR simply updates the FeathersVuexPagination component to use `h` and removes the legacy `createElement` function. A new release should mark the v1 version of the composition-api as required.